pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gha-geodata"
version = "0.1.0"
description = "GHA open geodata platform — ETL + API + CMS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Hillary Koros", email = "hkoros@icpac.net" },
]

dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "djangorestframework-gis>=1.0",
    "psycopg2-binary>=2.9",
    "httpx>=0.27",
    "geopandas>=0.14",
    "shapely>=2.0",
    "fiona>=1.9",
    "wagtail>=6.0",
    "wagtail-modeladmin>=2.0",
    "whitenoise",
    "gunicorn",
]

[project.optional-dependencies]
etl = [
    "click",
    "earthengine-api",
    "rasterio",
    "rioxarray",
    "xarray",
    "duckdb",
    "sqlalchemy",
    "geoalchemy2",
    "prometheus-client",
    "pyarrow",
]
gridded = [
    "zarr",
    "netcdf4",
    "numcodecs",
    "dask[complete]",
]
ibf = [
    "cdsapi>=0.7",
    "rasterstats>=0.19",
]
dev = ["pytest", "pytest-django", "ruff"]
scripts = ["pyarrow", "sqlalchemy", "geoalchemy2", "psycopg2-binary"]

[tool.hatch.build.targets.wheel]
packages = ["api", "etl"]
